Enhance onboarding experience with new components and FAQ section
