+++ components/onboarding/index.ts
@@ -1,0 +1,10 @@
+export { ActionButtons } from "./ActionButtons";
+export { HelpSection } from "./HelpSection";
+export { NavigationButtons } from "./NavigationButtons";
+export { OnboardingLayout } from "./OnboardingLayout";
+export { PromotionAlert } from "./PromotionAlert";
+export { Step } from "./Step";
+export { StepCard } from "./StepCard";
+export { Stepper } from "./Stepper";
+export { VideoTutorial } from "./VideoTutorial";
+export type { OnboardingStep } from "./types";
